+++ example/lib/di/injection.dart
@@ -1,4 +1,3 @@
-import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 import 'package:universal_navigation/di/u_nav_configurator.dart';
 import 'package:universal_navigation/navigation/bottom_navigation_page.dart';
@@ -7,20 +6,15 @@ import 'package:universal_navigation/navigation/core/tab_change_listener.dart';
 import 'package:universal_navigation_example/di/injection.config.dart';
 import 'package:universal_navigation_example/event_union.dart';
 
-final getIt = GetIt.instance;
-
 @injectableInit
 void configureInjection(String env) {
-  initUNavInjection<EventUnion>(getIt, env);
+  initUNavInjection<EventUnion>(env);
   $initGetIt(getIt, environment: env);
-  initUNavAppNavigatorInjection(getIt, env);
-  _initBottomNavigationInjection(getIt, env);
+  initUNavAppNavigatorInjection(env);
+  _initBottomNavigationInjection(env);
 }
 
-GetIt _initBottomNavigationInjection(GetIt get, String env) {
-  get.registerSingleton<BottomNavigationPage>(BottomNavigationPage(
-    get<TabChangeListener>(),
-    get<BottomNavigationBuilder>()
-  ));
-  return get;
+void _initBottomNavigationInjection(String env) {
+  getIt.registerSingleton<BottomNavigationPage>(BottomNavigationPage(
+      getIt<TabChangeListener>(), getIt<BottomNavigationBuilder>()));
 }