+++ example/lib/di/modules/bottom_nav_bridge_module.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:injectable/injectable.dart';
+import 'package:universal_navigation/di/u_nav_configurator.dart';
 import 'package:universal_navigation/navigation/core/bottom_navigation_builder.dart';
 import 'package:universal_navigation/navigation/models/navigation_flow_data/tabflow.dart';
 import 'package:universal_navigation/navigation/models/navigation_keys/bottom_nav_key.dart';
-import 'package:universal_navigation_example/di/injection.dart';
 import 'package:universal_navigation_example/pages/default_bottom_nav_bridge.dart';
 
 @module